Migrate utility functions for butlerUri into folder
Merge webdavutils.py into http.py for _butlerUri and simply move s3utils
into the folder for now. Someday we can perhaps make s3utils go away but
not today...
