+++ python/lsst/daf/butler/core/_butlerUri/__init__.py
@@ -1,5 +1,3 @@
 # Should only expose ButlerURI
 
 from ._butlerUri import ButlerURI
-__all__ = ("ButlerURI",)
-
