+++ python/lsst/daf/butler/core/_butlerUri/utils.py
@@ -28,6 +28,7 @@ import os
 
 from pathlib import Path, PurePath, PurePosixPath
 
+__all__ = ('os2posix', 'posix2os', 'NoTransaction')
 
 from typing import (
     Any,
