+++ tests/test_butler.py
@@ -66,9 +66,9 @@ from lsst.daf.butler import ButlerURI
 from lsst.daf.butler import script
 from lsst.daf.butler.registry import MissingCollectionError, OrphanedRecordError
 from lsst.daf.butler.core.repoRelocation import BUTLER_ROOT_TAG
-from lsst.daf.butler.core.s3utils import (setAwsEnvCredentials,
-                                          unsetAwsEnvCredentials)
-from lsst.daf.butler.core.webdavutils import isWebdavEndpoint
+from lsst.daf.butler.core._butlerUri.s3utils import (setAwsEnvCredentials,
+                                                     unsetAwsEnvCredentials)
+from lsst.daf.butler.core._butlerUri.http import isWebdavEndpoint
 
 from lsst.daf.butler.tests import MultiDetectorFormatter, MetricsExample
