+++ tests/test_uri.py
@@ -39,8 +39,8 @@ except ImportError:
         return cls
 
 from lsst.daf.butler import ButlerURI
-from lsst.daf.butler.core.s3utils import (setAwsEnvCredentials,
-                                          unsetAwsEnvCredentials)
+from lsst.daf.butler.core._butlerUri.s3utils import (setAwsEnvCredentials,
+                                                     unsetAwsEnvCredentials)
 
 TESTDIR = os.path.abspath(os.path.dirname(__file__))
